+++ docs/client/components/pages/Mention/CustomMentionEditor/index.js
@@ -1,22 +1,30 @@
 import React, { Component } from 'react';
 import { EditorState } from 'draft-js';
 import Editor from 'draft-js-plugins-editor';
-import createMentionPlugin from 'draft-js-mention-plugin';
+import createMentionPlugin, { defaultSuggestionsFilter } from 'draft-js-mention-plugin';
 import editorStyles from './editorStyles.css';
 import mentions from './mentions';
 
-const mentionPlugin = createMentionPlugin({ mentions });
+const mentionPlugin = createMentionPlugin({ mentions, entityMutability: 'IMMUTABLE' });
+const { SearchSuggestions } = mentionPlugin;
 const plugins = [mentionPlugin];
 
 export default class CustomMentionEditor extends Component {
 
   state = {
     editorState: EditorState.createEmpty(),
+    suggestions: mentions,
   };
 
   onChange = (editorState) => {
     this.setState({
       editorState,
+    });
+  };
+
+  onSearchChange = ({ value }) => {
+    this.setState({
+      suggestions: defaultSuggestionsFilter(value, mentions),
     });
   };
 
@@ -28,10 +36,14 @@ export default class CustomMentionEditor extends Component {
     return (
       <div className={ editorStyles.editor } onClick={ this.focus }>
         <Editor
-          editorState={this.state.editorState}
+          editorState={ this.state.editorState }
           onChange={this.onChange}
           plugins={plugins}
           ref="editor"
+        />
+        <SearchSuggestions
+          onSearchChange={ this.onSearchChange }
+          suggestions={ this.state.suggestions }
         />
       </div>
     );