+++ docs/client/components/pages/Undo/index.js
@@ -32,7 +32,8 @@ export default class App extends Component {
         </Container>
         <AlternateContainer>
           <Heading level={ 2 }>Getting Started</Heading>
-          <Code code="npm install draft-js-history-plugin --save" />
+          <Code code="npm install draft-js-plugins-editor@1.0.0-beta1 --save" />
+          <Code code="npm install draft-js-history-plugin@1.0.0-beta1 --save" />
           <Code code={ gettingStarted } name="gettingStarted.js" />
           <Heading level={ 3 }>Importing the default styles</Heading>
           <p>