+++ draft-js-counter-plugin/src/LineCounter/index.js
@@ -22,7 +22,7 @@ class LineCounter extends Component {
   }
 
   render() {
-    const { editorState, limit, ...props } = this.props; // eslint-disable-line no-use-before-define
+    const { editorState, limit } = this.props;
     const count = this.getLineCount(editorState);
     const classNames = this.getClassNames(count, limit);