+++ draft-js-mention-plugin/src/SearchSuggestions/MentionOption/Avatar/index.js
@@ -3,7 +3,11 @@ import React from 'react';
 const Avatar = ({ mention, theme }) => {
   if (mention.has('avatar')) {
     return (
-      <img src={ mention.get('avatar') } className={ theme.get('autocompleteEntryAvatar') } />
+      <img
+        src={ mention.get('avatar') }
+        className={ theme.get('autocompleteEntryAvatar') }
+        role="presentation"
+      />
     );
   }