+++ draft-js-sticker-plugin/src/Sticker/index.js
@@ -36,6 +36,7 @@ export default class Sticker extends Component {
         <img
           className={ theme.get('stickerImage') }
           src={ stickers.getIn(['data', data.id, 'url']) }
+          role="presentation"
         />
         { this.props.attachRemoveButton ? removeButton : null }
       </figure>